apiserver: add tests and #1582 for security groups.
Implements #1582 for security groups.
We now sort by the group name instead of by id.

+++ internal/handlers/security_group.go
@@ -95,7 +95,7 @@ func (api *API) ListSecurityGroups(c *gin.Context) {
 	api.sendList(c, ctx, func(db *gorm.DB) (fetchmgr.ResourceList, error) {
 		var items securityGroupList
 		db = api.SecurityGroupIsReadableByCurrentUser(c, db)
-		db = FilterAndPaginateWithQuery(db, &models.SecurityGroup{}, c, query, "id")
+		db = FilterAndPaginateWithQuery(db, &models.SecurityGroup{}, c, query, "group_name")
 		result := db.Find(&items)
 		if result.Error != nil && !errors.Is(result.Error, gorm.ErrRecordNotFound) {
 			return nil, result.Error
@@ -290,13 +290,9 @@ func (api *API) CreateSecurityGroup(c *gin.Context) {
 			OutboundRules:    request.OutboundRules,
 			GroupDescription: request.GroupDescription,
 		}
-		if res := tx.Create(&sg); res.Error != nil {
-			return res.Error
-		}
-
-		// Replace the organization's SecurityGroupId field
-		org.SecurityGroupId = sg.ID
-		if res := tx.Save(&org); res.Error != nil {
+		if res := tx.
+			Clauses(clause.Returning{Columns: []clause.Column{{Name: "revision"}}}).
+			Create(&sg); res.Error != nil {
 			return res.Error
 		}
 
@@ -367,12 +363,15 @@ func (api *API) DeleteSecurityGroup(c *gin.Context) {
 	}
 
 	sg := models.SecurityGroup{}
-
 	err = api.transaction(ctx, func(tx *gorm.DB) error {
 
 		if res := api.SecurityGroupIsWriteableByCurrentUser(c, tx).
 			First(&sg, "id = ?", secGroupID); res.Error != nil {
 			return NewApiResponseError(http.StatusNotFound, models.NewNotFoundError("vpc"))
+		}
+
+		if sg.ID == sg.OrganizationId {
+			return NewApiResponseError(http.StatusBadRequest, models.NewNotAllowedError("default security group cannot be deleted"))
 		}
 
 		if res := tx.Delete(&sg, "id = ?", sg.ID); res.Error != nil {
